+++ client/video/run.ts
@@ -86,7 +86,7 @@ async function init(config) {
 }
 
 async function main(config, objects) {
-  const video = document.getElementById('video');
+  const video = document.getElementById('video') as HTMLVideoElement;
   const t0 = performance.now();
   objects.detected = {};
   objects.classified = {};
@@ -142,7 +142,6 @@ async function main(config, objects) {
   if (!el) return;
   if (config.ui.text) {
     let msg = '';
-    // @ts-ignore
     msg += `camera: ${video?.videoWidth} x ${video?.videoHeight} processing: ${input.width} x ${input.height}<br>`;
     if (objects.human.length > 0) msg += `human: ${log.str([...objects.human])}<br>`;
     for (const [key, val] of Object.entries(objects.detected)) {
@@ -156,8 +155,7 @@ async function main(config, objects) {
   } else {
     el.innerHTML = '';
   }
-  // @ts-ignore
-  document.getElementById('status').innerText = '';
+  (document.getElementById('status') as HTMLElement).innerText = '';
   objects.perf['Total time'] = Math.trunc(t1 - t0);
   const mem = tf.memory();
   objects.perf['System Memory'] = Math.trunc(mem.numBytes / 1024 / 1024).toLocaleString();