+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Repository: **<git+https://github.com/vladmandic/pigallery.git>**
 
 ### **HEAD -> master** 2021/04/23 mandic00@live.com
 
+- refactoring index and details for typescript
+- additional typescript refactoring
+- typescript conversion for process
+- refactor shared modules
+- add score filter on image display
 
 ### **origin/master** 2021/04/22 mandic00@live.com